content/media/video/test/test_start.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: start</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v1'></video><audio id='a1'></audio>
<pre id="test">
<script class="testbody" type="text/javascript">
var v1 = document.getElementById('v1');
var a1 = document.getElementById('a1');

is(v1.start, 0);
is(a1.start, 0);

v1.start = 2.5;
a1.start = 2.5;
is(v1.start, 2.5);
is(a1.start, 2.5);
</script>
</pre>
</body>
</html>
